\name{lrmodels-package}
\alias{lrmodels-package}
\alias{lrmodels}
\docType{package}
\title{
\packageTitle{lrmodels}
}
\description{
\packageDescription{lrmodels}
}
\details{
\packageDescription{lrmodels} is an R package designed for handling linear regression models. It provides a set of tools for linear regression modeling and analysis. The package includes functions for model fitting, summary statistics, visualization, and diagnostic plots, making it a valuable resource for data analysts and researchers working with linear regression.

This package is particularly useful for tasks such as:
- Fitting linear regression models to your data.
- Extracting and interpreting regression coefficients.
- Creating diagnostic plots to assess model assumptions.
- Calculating residuals and predicted values.
- Summarizing regression model results.

For more details on how to use the package and its functions, please refer to the package documentation and the examples provided in the vignettes.

}

\author{
\packageAuthor{lrmodels}

Maintainer: \packageMaintainer{lrmodels}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[lrmodels:lrmodels-package]{lrmodels}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
