\name{lrmodels-package}
\alias{lrmodels-package}
\alias{lrmodels}
\docType{package}
\title{
\packageTitle{lrmodels}
}
\description{
\packageDescription{lrmodels}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lrmodels}
\packageIndices{lrmodels}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{lrmodels}

Maintainer: \packageMaintainer{lrmodels}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[lrmodels:lrmodels-package]{lrmodels}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
