\name{lrmodels-package}
\alias{lrmodels-package}
\alias{lrmodels}
\docType{package}
\title{lrmodels: Linear Regression Models in R}

\description{
The \pkg{lrmodels} package is designed for handling linear regression models. It provides a set of tools for linear regression modeling and analysis. The package includes functions for model fitting, summary statistics, visualization, and diagnostic plots, making it a valuable resource for data analysts and researchers working with linear regression.

This package is particularly useful for tasks such as:
\itemize{
  \item Fitting linear regression models to your data.
  \item Extracting and interpreting regression coefficients.
  \item Creating diagnostic plots to assess model assumptions.
  \item Calculating residuals and predicted values.
  \item Summarizing regression model results.
}

For more details on how to use the package and its functions, please refer to the package documentation and the examples provided in the vignettes.
}

\author{
\packageAuthor{lrmodels}

Maintainer: \packageMaintainer{lrmodels}
}

\references{
~~ Literature or other references for background information ~~
}

\keyword{package}

\seealso{
Optional links to other man pages, e.g. \code{\link[lrmodels:lrmodels-package]{lrmodels}}
}

